usb: add USB 2.1 Binary Object Store support
Add support for USB BOS descriptors, which store a list of device-level
capability descriptors that can be used to declare support for extended
features such as Microsoft OS 2.0 descriptors or WebUSB.

BOS descriptors are technically part of the USB 3.0 specification.
Devices that support BOS descriptors while otherwise remaining USB 2.0
generally advertise their version as 2.1 to reflect this in-between mix.